prefect  the drop function
implement the restriction that once the capsule collide the other object
except the NPC BOX, the drop function will disable.
the drop will clear the selected thing in the inventory, and the pick
will put them all in the inventory.
but the drop disable nodification is still hard to implement.